Moved auth to endpoints.
